+++ src/components/ui/dropdown/Dropdown.tsx
@@ -1,35 +1,42 @@
 import DropdownList from '@/components/ui/dropdown/DropdownList';
+import { DownArrowIcon } from '@/components/ui/icons/ArrowIcon';
 import DropdownIcon from '@/components/ui/icons/DropdownIcon';
-import LabeledContent from '@/components/ui/LabeledContent';
+import useClickOutsideRef from '@/hooks/useClickOutsideRef';
+import { useState } from 'react';
 
 interface PDropdown {
-  label: string;
   placeholder: string;
   options: string[];
-  selected: string[];
-  setSelected: (selected: string) => void;
-  isOpen: boolean;
-  handleOpen: () => void;
+  selected: string[] | string;
+  setSelected: (value: string) => void;
+  id?: string;
+  color?: 'primary' | 'gray';
 }
 
-const Dropdown: React.FC<PDropdown> = ({ label, placeholder, options, selected, setSelected, isOpen, handleOpen }) => {
+const Dropdown: React.FC<PDropdown> = ({ placeholder, options, selected, setSelected, id, color = 'primary' }) => {
+  const [isOpen, setIsOpen] = useState<boolean>(false);
+
+  const ref = useClickOutsideRef<HTMLDivElement>(() => setIsOpen(false));
+
+  const isUnselected = (typeof selected === 'string' && selected === '') || selected.length === 0;
+
   return (
-    <div className='relative h-10 w-full pl-5 pr-2'>
-      <button className='flex h-full w-full items-center justify-between text-left' onClick={handleOpen}>
-        <LabeledContent
-          label={label}
-          icon={
-            <div>
-              <DropdownIcon />
-            </div>
-          }
-        >
-          <span className={`${selected.length === 0 ? 'text-disabled' : 'text-black'} line-clamp-1 text-body1`}>
-            {selected.length === 0 ? placeholder : selected.length === options.length ? '전체' : selected.join(', ')}
-          </span>
-        </LabeledContent>
+    <div className='relative w-full' ref={ref}>
+      <button
+        id={id}
+        className='flex w-full items-center gap-2 rounded-sm border border-input px-4 py-2.5 text-body2 focus:outline-activate'
+        onClick={() => setIsOpen((prev) => !prev)}
+      >
+        <p className={`line-clamp-1 flex-grow text-start ${isUnselected ? 'text-disabled' : ''}`}>
+          {isUnselected ? placeholder : typeof selected === 'string' ? selected : selected.join(', ')}
+        </p>
+        <div>{color === 'primary' ? <DropdownIcon /> : <DownArrowIcon size={22} color='#9e9e9e' />}</div>
       </button>
-      {isOpen && <DropdownList options={options} selected={selected} setSelected={setSelected} />}
+      {isOpen && (
+        <div className='absolute left-0 top-10 w-full'>
+          <DropdownList options={options} selected={selected} setSelected={setSelected} />
+        </div>
+      )}
     </div>
   );
 };
